homework 15-message system - added corrections
