+++ hw15-message.system/src/main/resources/application.yml
@@ -7,3 +7,7 @@ spring:
     url: jdbc:h2:mem:testDB;DB_CLOSE_DELAY=-1
     data-username: sa
     driver-class-name: org.h2.Driver
+
+message-address:
+  frontAddress: Frontend
+  dbAddress: DB
